I lied in the last check-in. The black triangle thing wasn't fixed. It is now. Here's what I learned:
Observation: Immediately when the first puzzle's start fold animation finished, I'd grab the fold. Where the screenshot should be presented in the body of the page, instead the translucent page graphic would show.
Assumption: The translucent page graphic was being loaded at the same time the screenshot graphic was being loaded.
Confusion: The manual says each thread maintains a separate context, but it seems that some part of it doesn't work right. Could it be that the texture loader isn't thread safe?
Work-around: I put a semaphore around all code that makes OpenGL calls. Maybe it's just me worrying, but I think it does affect performance slightly. But it fixes the problem.

git-svn-id: file:///Users/brent/Dropbox/Development/Repositories/ZenSudoku/trunk@144 4be4153c-cb45-4615-a51f-74384bbadf62